Windows build fixes

+++ src/sbs/globals.cpp
@@ -309,7 +309,7 @@ std::string ToString(double number)
 	return buffer;
 }
 
-std::string ToString(size_t number)
+std::string ToString(unsigned int number)
 {
 	char buffer[50];
 #if defined(__VISUALC__)
@@ -320,13 +320,13 @@ std::string ToString(size_t number)
 	return buffer;
 }
 
-std::string ToString(unsigned int number)
+std::string ToString(unsigned long number)
 {
 	char buffer[50];
 #if defined(__VISUALC__)
-	_snprintf_s(buffer, sizeof(buffer), 13, "%d", number);
+	_snprintf_s(buffer, sizeof(buffer), 13, "%lu", number);
 #else
-	snprintf(buffer, sizeof(buffer), "%d", number);
+	snprintf(buffer, sizeof(buffer), "%lu", number);
 #endif
 	return buffer;
 }
